+++ R/api-progress.R
@@ -3,8 +3,7 @@
 #'   (2) were being built if the last [make()] quit unexpectedly.
 #' @description Similar to [progress()].
 #' @seealso [diagnose()], [drake_get_session_info()],
-#'   [built()], [imported()],
-#'   [readd()], [drake_plan()], [make()]
+#'   [cached()], [readd()], [drake_plan()], [make()]
 #' @export
 #' @return A character vector of target names.
 #' @inheritParams cached
@@ -31,8 +30,7 @@ in_progress <- function(path = getwd(), search = TRUE,
 #' [diagnose()] should eliminate the strict need
 #' for ordinary error messages printed to the console.
 #' @seealso [diagnose()], [drake_get_session_info()],
-#'   [built()], [imported()],
-#'   [readd()], [drake_plan()], [make()]
+#'   [cached()], [readd()], [drake_plan()], [make()]
 #' @export
 #' @return A character vector of target names.
 #' @inheritParams cached
@@ -68,8 +66,7 @@ failed <- function(path = getwd(), search = TRUE,
 #' to build are listed as `"finished"` or `"in progress"`.
 #' Skipped objects are not listed.
 #' @seealso [diagnose()], [drake_get_session_info()],
-#'   [built()], [imported()],
-#'   [readd()], [drake_plan()], [make()]
+#'   [cached()], [readd()], [drake_plan()], [make()]
 #' @export
 #'
 #' @return The build progress of each target reached by