+++ R/utils-deprecate.R
@@ -891,6 +891,86 @@ drake_strings <- function(...) {
   out
 }
 
+#' @title Deprecated. List all the built targets (non-imports) in the cache.
+#' @description Deprecated on 2019-01-08.
+#' @details Targets are listed in the workflow plan
+#' data frame (see [drake_plan()].
+#' @seealso [cached()], [loadd()]
+#' @export
+#' @return Character vector naming the built targets in the cache.
+#' @inheritParams cached
+#' @param jobs number of jobs/workers for parallel processing
+#' @examples
+#' \dontrun{
+#' test_with_dir("Quarantine side effects.", {
+#' load_mtcars_example() # Load drake's canonical example.
+#' make(my_plan) # Run the project, build all the targets.
+#' built() # List all the cached targets (built objects and files).
+#' # For file targets, only the fingerprints/hashes are stored.
+#' })
+#' }
+built <- function(
+  path = getwd(), search = TRUE,
+  cache = drake::get_cache(path = path, search = search, verbose = verbose),
+  verbose = 1L,
+  jobs = 1
+) {
+  .Deprecated(
+    "built",
+    package = "drake",
+    msg = paste("built() is deprecated.",
+                "Use cached(no_imported_objects = TRUE)) instead.")
+  )
+  if (is.null(cache)) {
+    return(character(0))
+  }
+  out <- cache$list(namespace = cache$default_namespace)
+  out <- parallel_filter(
+    out,
+    f = function(target) {
+      !is_imported_cache(target = target, cache = cache)
+    },
+    jobs = jobs
+  )
+  display_keys(out)
+}
+
+#' @title Deprecated. Search up the file system
+#'   for the nearest root path of a drake project.
+#' @description Deprecated on 2019-01-08.
+#' @details Only works if the cache is a file system
+#' in a folder named `.drake` (default).
+#' @export
+#' @seealso [drake_plan()], [make()]
+#' @return File path of the nearest drake project or `NULL`
+#'   if no drake project is found.
+#' @param path starting path for search back for the project.
+#'   Should be a subdirectory of the drake project.
+#' @examples
+#' \dontrun{
+#' test_with_dir("Quarantine side effects.", {
+#' load_mtcars_example() # Get the code with drake_example("mtcars").
+#' make(my_plan) # Run the project, build the target.
+#' # Find the root directory of the current drake project.
+#' # Search up through parent directories if necessary.
+#' find_cache()
+#' })
+#' }
+find_project <- function(path = getwd()) {
+  .Deprecated(
+    "find_project",
+    package = "drake",
+    msg = paste("find_project() is deprecated.",
+                "Use find_cache() instead.")
+  )
+
+  cache <- find_cache(path = path)
+  if (is.null(cache)) {
+    return(NULL)
+  }
+  return(dirname(cache))
+}
+
 #' @title Deprecated
 #' @description 2019-01-03
 #' @export
@@ -1225,6 +1305,64 @@ read_drake_plan <- function(
     )
   )
   drake_plan()
+}
+
+#' @title Deprecated. List all the imports in the drake cache.
+#' @description Deprecated on 2019-01-08.
+#' @details An import is a non-target object processed
+#' by [make()]. Targets in the workflow
+#' plan data frame (see [drake_config()]
+#' may depend on imports.
+#' @seealso [cached()], [loadd()]
+#' @export
+#' @return Character vector naming the imports in the cache.
+#'
+#' @inheritParams cached
+#'
+#' @param files_only logical, whether to show imported files only
+#'   and ignore imported objects. Since all your functions and
+#'   all their global variables are imported, the full list of
+#'   imported objects could get really cumbersome.
+#'
+#' @param jobs number of jobs/workers for parallel processing
+#'
+#' @examples
+#' \dontrun{
+#' test_with_dir("Quarantine side effects.", {
+#' load_mtcars_example() # Load the canonical example.
+#' make(my_plan) # Run the project, build the targets.
+#' imported() # List all the imported objects/files in the cache.
+#' # For imported files, only the fingerprints/hashes are stored.
+#' })
+#' }
+imported <- function(
+  files_only = FALSE, path = getwd(), search = TRUE,
+  cache = drake::get_cache(path = path, search = search, verbose = verbose),
+  verbose = 1L,
+  jobs = 1
+) {
+  .Deprecated(
+    "imported",
+    package = "drake",
+    msg = paste(
+      "imported() is deprecated.",
+      "Use setdiff(cached(), cached(no_imported_objects = TRUE)) instead."
+    )
+  )
+  if (is.null(cache)) {
+    return(character(0))
+  }
+  targets <- cache$list(namespace = cache$default_namespace)
+  targets <- parallel_filter(
+    targets,
+    f = function(target) {
+      is_imported_cache(target = target, cache = cache)
+    },
+    jobs = jobs
+  )
+  if (files_only)
+    targets <- parallel_filter(targets, f = is_encoded_path, jobs = jobs)
+  display_keys(targets)
 }
 
 #' @title deprecated